# docs/planning/openapi.yaml
openapi: 3.0.3
info:
  title: HelpHut Food Rescue Management
  description: >
    Centralized platform for coordinating food donations, volunteers, and partner organizations.
    Aligns with the HelpHut mission to efficiently rescue food in Austin, reduce waste, and help those in need.
    This specification is maintained under a TDD approach, with auto-generated tests ensuring coverage of each endpoint.
  version: "1.0.0"
servers:
  - url: https://api.helphut.org/v1
security:
  - BearerAuth: []
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: List tickets
      operationId: listTickets
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by ticket status
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          required: false
          description: Filter by ticket priority
          schema:
            $ref: "#/components/schemas/TicketPriority"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Array of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Tickets]
      summary: Create a new ticket
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketCreate"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: Retrieve a ticket by ID
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "200":
          description: A ticket object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Tickets]
      summary: Update a ticket by ID
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Tickets]
      summary: Delete a ticket by ID
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "204":
          description: Deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathUUID:
      name: id
      in: path
      required: true
      description: UUID of the resource
      schema:
        type: string
        format: uuid
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 10
      description: Maximum number of items to return
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip before starting to collect results
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Detailed error description
...
    TicketStatus:
      type: string
      enum: [Submitted, Scheduled, InTransit, Delivered, Completed]
    TicketPriority:
      type: string
      enum: [Urgent, Routine]
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        donation_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        volunteer_id:
          type: string
          format: uuid
        partner_org_id:
          type: string
          format: uuid
        pickup_location_id:
          type: string
          format: uuid
        dropoff_location_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TicketCreate:
      type: object
      properties:
        donation_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        volunteer_id:
          type: string
          format: uuid
        partner_org_id:
          type: string
          format: uuid
        pickup_location_id:
          type: string
          format: uuid
        dropoff_location_id:
          type: string
          format: uuid
    TicketUpdate:
      type: object
      properties:
        donation_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        volunteer_id:
          type: string
          format: uuid
        partner_org_id:
          type: string
          format: uuid
        pickup_location_id:
          type: string
          format: uuid
        dropoff_location_id:
          type: string
          format: uuid
...